Ajout Fonction Conversion degré vers rapport PWM
Plus ajouts variables define (Pour le CAN)

+++ dsPic30f2010-EmulateurAX12.X (Version 4Mhz + PLLx16)/CAN10bits.h
@@ -15,7 +15,7 @@ extern "C" {
 #include "Include_Library.h"
     
     void CAN_init();
-    unsigned int LectureTensionMoteur();
+    float LectureTensionMoteur();
 
 
 #ifdef	__cplusplus
